chore: remove circular imports
This project uses circular imports, which cannot be bundled
using `node-bundle`. Get rid of the circular import.
