+++ lib/private/handlers/index.ts
@@ -1,4 +1,3 @@
-import type { AwsDestination } from '@aws-cdk/cloud-assembly-schema';
 import { ContainerImageAssetHandler } from './container-images';
 import { FileAssetHandler } from './files';
 import {
@@ -7,7 +6,6 @@ import {
   FileManifestEntry,
   type IManifestEntry,
 } from '../../asset-manifest';
-import type { ClientOptions } from '../../aws';
 import type { IAssetHandler, IHandlerHost, IHandlerOptions } from '../asset-handler';
 
 export function makeAssetHandler(
@@ -24,13 +22,4 @@ export function makeAssetHandler(
   }
 
   throw new Error(`Unrecognized asset type: '${asset}'`);
-}
-
-export function destinationToClientOptions(destination: AwsDestination): ClientOptions {
-  return {
-    assumeRoleArn: destination.assumeRoleArn,
-    assumeRoleExternalId: destination.assumeRoleExternalId,
-    assumeRoleAdditionalOptions: destination.assumeRoleAdditionalOptions,
-    region: destination.region,
-  };
 }
