+++ package.json
@@ -56,6 +56,7 @@
     "jest": "^29.7.0",
     "jest-junit": "^15",
     "jszip": "^3.10.1",
+    "madge": "^8.0.0",
     "prettier": "^3.4.2",
     "projen": "^0.90.6",
     "ts-jest": "^29.2.5",
